% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{american_to_implied}
\alias{american_to_implied}
\title{Convert American Odds to Implied Probability}
\usage{
american_to_implied(american_odds)
}
\arguments{
\item{american_odds}{Numeric vector of American odds}
}
\value{
Numeric vector of implied probabilities
}
\description{
Convert American Odds to Implied Probability
}
\examples{
american_to_implied(100)
}
