% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_consensus_odds.R
\name{calculate_consensus_odds}
\alias{calculate_consensus_odds}
\title{Calculate Consensus Odds}
\usage{
calculate_consensus_odds(odds_df, min_bookmakers = 3)
}
\arguments{
\item{odds_df}{Data frame from get_nfl_odds}

\item{min_bookmakers}{Minimum number of bookmakers required}
}
\value{
Data frame with consensus odds
}
\description{
Calculate Consensus Odds
}
