% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_means.R
\name{calculate_means}
\alias{calculate_means}
\title{Calculate Means by Team History}
\usage{
calculate_means(data)
}
\arguments{
\item{data}{Data frame containing NFL statistics}
}
\value{
Data frame with calculated means
}
\description{
Calculate Means by Team History
}
