% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{clean_team_name}
\alias{clean_team_name}
\title{Function to clean team names using the mapping}
\usage{
clean_team_name(team_name)
}
\arguments{
\item{team_name}{Character string of team abbreviation or name}
}
\value{
Character string of cleaned team name
}
\description{
Function to clean team names using the mapping
}
