% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare Multiple Model Results}
\usage{
compare_models(..., round_digits = 3)
}
\arguments{
\item{...}{Named list of model CV results}

\item{round_digits}{Number of digits to round metrics (default = 3)}
}
\value{
A data frame comparing model metrics
}
\description{
Compare Multiple Model Results
}
