% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injury_analysis.R
\name{get_team_injury_report}
\alias{get_team_injury_report}
\title{Get injury report for a specific team & week}
\usage{
get_team_injury_report(team_abbr, season, week, injury_data)
}
\description{
Get injury report for a specific team & week
}
