% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{implied_to_american}
\alias{implied_to_american}
\title{Convert Implied Probability to American Odds}
\usage{
implied_to_american(implied_prob)
}
\arguments{
\item{implied_prob}{Numeric vector of probabilities (0-1)}
}
\value{
Numeric vector of American odds
}
\description{
Convert Implied Probability to American Odds
}
\examples{
implied_to_american(0.5)
}
