% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{last_3_or_season_mean}
\alias{last_3_or_season_mean}
\title{Calculate Last 3 Games Average}
\usage{
last_3_or_season_mean(data, n = 3)
}
\arguments{
\item{data}{Numeric vector containing statistics}

\item{n}{Number of games to look back (default 3)}
}
\value{
Numeric mean value
}
\description{
Calculate Last 3 Games Average
}
\keyword{Calculate}
\keyword{\code{n}}
\keyword{are}
\keyword{available.}
\keyword{entire}
\keyword{fewer}
\keyword{games}
\keyword{internal}
\keyword{last}
\keyword{mean}
\keyword{of}
\keyword{or}
\keyword{than}
\keyword{the}
\keyword{vector}
\keyword{when}
