% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{map_team_abbreviation}
\alias{map_team_abbreviation}
\title{Map Team Abbreviations (single source of truth)}
\usage{
map_team_abbreviation(abbr)
}
\arguments{
\item{abbr}{character vector of team abbreviations}
}
\description{
Map Team Abbreviations (single source of truth)
}
\keyword{internal}
