% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfl_schedule_2024}
\alias{nfl_schedule_2024}
\title{NFL Schedule 2024}
\format{
A data frame with rows and columns:
\describe{
\item{match_number}{Game of the season}
\item{away_team}{Three-letter abbreviation of away team}
\item{home_team}{Three-letter abbreviation of home team}
\item{Date}{Date of the game}
\item{Location}{Location of Game}
\item{week}{Week game was played}
}
}
\source{
{https://fixturedownload.com/download/csv/nfl-2024}
}
\usage{
nfl_schedule_2024
}
\description{
A dataset containing the NFL schedule for the 2024 season
}
\keyword{datasets}
