% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NFL_TEAM_MAPPINGS}
\alias{NFL_TEAM_MAPPINGS}
\title{NFL Team Name Mappings}
\format{
A named character vector
}
\usage{
NFL_TEAM_MAPPINGS
}
\description{
A named vector mapping team abbreviations to full team names
Used for converting between ESPN abbreviations and full team names
}
\keyword{datasets}
