% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_weekly.R
\name{prepare_weekly}
\alias{prepare_weekly}
\title{Process NFL Weekly Data}
\usage{
prepare_weekly(years)
}
\arguments{
\item{years}{Numeric vector of years to process}
}
\value{
A data frame of processed weekly NFL data
}
\description{
Process NFL Weekly Data
}
\examples{
\dontrun{
weekly_data <- prepare_weekly(2024)
}
}
