% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{print.formula_evaluation}
\alias{print.formula_evaluation}
\title{Print method for formula evaluation}
\usage{
\method{print}{formula_evaluation}(x, digits = 3)
}
\arguments{
\item{x}{Formula evaluation object}

\item{digits}{Number of digits to display}
}
\description{
Print method for formula evaluation
}
