% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_cv.R
\name{print.model_cv}
\alias{print.model_cv}
\title{Print method for CV results}
\usage{
\method{print}{model_cv}(x, ...)
}
\arguments{
\item{x}{CV results object}

\item{...}{Additional arguments}
}
\description{
Print method for CV results
}
