% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nfl_schedule.R
\name{print.nfl_schedule}
\alias{print.nfl_schedule}
\title{Print method for nfl_schedule objects}
\usage{
\method{print}{nfl_schedule}(x, ...)
}
\arguments{
\item{x}{An object of class nfl_schedule}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for nfl_schedule objects
}
