% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R, R/update_weekly.R
\name{update_weekly}
\alias{update_weekly}
\title{Update NFL Weekly Data}
\usage{
update_weekly(years, replace_existing = FALSE)

update_weekly(years, replace_existing = FALSE)
}
\arguments{
\item{years}{Vector of years to update}

\item{replace_existing}{Logical, whether to replace existing years' data,
must be true to load data before 2015}
}
\value{
Updated weekly dataset

Updated weekly dataset
}
\description{
Update NFL Weekly Data

Update NFL Weekly Data
}
