% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{last_3_or_season_mean}
\alias{last_3_or_season_mean}
\title{Calculate Last 3 Games Average}
\usage{
last_3_or_season_mean(data, n = 3)
}
\arguments{
\item{data}{Data frame containing stats}

\item{n}{Number of games to look back}
}
\value{
Numeric mean value
}
\description{
Calculate Last 3 Games Average
}
\keyword{3}
\keyword{Calculate}
\keyword{games}
\keyword{internal}
\keyword{last}
\keyword{mean}
\keyword{of}
\keyword{or}
\keyword{season}
