% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{last_3_or_season_mean}
\alias{last_3_or_season_mean}
\title{Calculate Last 3 Games Average}
\usage{
last_3_or_season_mean(data, n = 3)
}
\arguments{
\item{data}{Numeric vector containing statistics}

\item{n}{Number of games to look back}
}
\value{
Numeric mean value
}
\description{
Calculate mean of last \code{n} games or of the entire vector when fewer than \code{n} games are available
}
\keyword{3}
\keyword{Calculate}
\keyword{games}
\keyword{internal}
\keyword{last}
\keyword{mean}
\keyword{of}
\keyword{or}
\keyword{season}
