% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{map_team_abbreviation}
\alias{map_team_abbreviation}
\title{Map Team Abbreviations}
\usage{
map_team_abbreviation(abbr)
}
\arguments{
\item{abbr}{NFL old team name}
}
\description{
Map Team Abbreviations
}
\keyword{internal}
