% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prediction.R
\name{prepare_predictions}
\alias{prepare_predictions}
\title{Prepare NFL Prediction Data}
\usage{
prepare_predictions(weekly_data)
}
\arguments{
\item{api_key}{Character string containing your API key for the-odds-api.com}
}
\value{
A data frame of prepared prediction data
}
\description{
Prepares NFL game data with team statistics and consensus odds for prediction.
}
