% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R, R/update_games.R
\name{update_games}
\alias{update_games}
\title{Update NFL Game Data}
\usage{
update_games(years, replace_existing = FALSE)

update_games(years, replace_existing = FALSE)
}
\arguments{
\item{years}{Vector of years to update}

\item{replace_existing}{Logical, whether to replace existing years' data}
}
\value{
Updated game dataset

Updated game dataset
}
\description{
Update NFL Game Data

Update NFL Game Data
}
